First significant refactoring to organize code
